# s_firebase/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Firebase Authentication API",
      "version": "1.0.0",
      "description": "API de autenticação que permite login e registro via email/password ou OAuth (Google/Facebook)."
    },
    "servers": [
      {
        "url": "http://localhost:5001",
        "description": "Servidor Local"
      }
    ],
    "paths": {
      "/v1/auth/register": {
        "post": {
          "summary": "Regista um utilizador",
          "description": "Regista um utilizador com email/password ou via OAuth.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "email": { "type": "string", "example": "user@example.com" },
                        "password": { "type": "string", "example": "123456" },
                        "role_code": { "type": "string", "example": "ADMIN123" }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "oauth_provider": { "type": "string", "example": "google" },
                        "token": { "type": "string", "example": "google_oauth_token" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Utilizador registado com sucesso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user_id": { "type": "string", "example": "uid123" },
                      "role": { "type": "string", "example": "admin" }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Erro no pedido",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "string", "example": "Invalid request data" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/v1/auth/login": {
        "post": {
          "summary": "Login de utilizador",
          "description": "Faz login com email/password ou OAuth (Google/Facebook).",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "email": { "type": "string", "example": "user@example.com" },
                        "password": { "type": "string", "example": "123456" }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "oauth_provider": { "type": "string", "example": "facebook" },
                        "token": { "type": "string", "example": "facebook_oauth_token" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login bem-sucedido",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user_id": { "type": "string", "example": "uid123" },
                      "role": { "type": "string", "example": "user" }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Credenciais inválidas",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "string", "example": "Invalid credentials" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/v1/auth/user/{user_id}": {
        "get": {
          "summary": "Obtém informações do utilizador",
          "description": "Obtém detalhes do utilizador autenticado.",
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Dados do utilizador",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user_id": { "type": "string", "example": "uid123" },
                      "email": { "type": "string", "example": "user@example.com" },
                      "role": { "type": "string", "example": "admin" }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Utilizador não encontrado",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "string", "example": "User not found" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/v1/auth/verify-token": {
        "get": {
          "summary": "Verifica um token JWT",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Token válido",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user_id": { "type": "string", "example": "uid123" },
                      "role": { "type": "string", "example": "user" }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Token inválido ou expirado",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "string", "example": "Invalid or expired token" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
  